Bug fix premove
Updated Board.validMove() to have a third parameter which was the enemies on the board. This ensures that you don't move into an enemy. Because that would not be a valid move.

Note: this check is entirely used by the NPC classes, the player can still blunder into an Enemy if they want.