Restored Javadoc/Refractored for ID
In order to get the replay to work a pass by reference system was needed for the recorder getting actors to be moved. Thus action record needed to store the id of an actor rather than the object itself. I also restored my old javadoc comments which were strangely deleted and added a couple new ones.